micromegas_perfetto::protos::c_switch_etw_event

Enum OldThreadWaitReason

#[repr(i32)]
pub enum OldThreadWaitReason {

    Executive = 0,
    FreePage = 1,
    PageIn = 2,
    PoolAllocation = 3,
    DelayExecution = 4,
    Suspend = 5,
    UserRequest = 6,
    WrExecutive = 7,
    WrFreePage = 8,
    WrPageIn = 9,
    WrPoolAllocation = 10,
    WrDelayExecution = 11,
    WrSuspended = 12,
    WrUserRequest = 13,
    WrEventPair = 14,
    WrQueue = 15,
    WrLpcReceiver = 16,
    WrLpcReply = 17,
    WrVirtualMemory = 18,
    WrPageOut = 19,
    WrRendezVous = 20,
    WrKeyedEvent = 21,
    WrTerminated = 22,
    WrProcessInSwap = 23,
    WrCpuRateControl = 24,
    WrCalloutStack = 25,
    WrKernel = 26,
    WrResource = 27,
    WrPushLock = 28,
    WrMutex = 29,
    WrQuantumEnd = 30,
    WrDispatchInt = 31,
    WrPreempted = 32,
    WrYieldExecution = 33,
    WrFastMutex = 34,
    WrGuardMutex = 35,
    WrRundown = 36,
    MaximumWaitReason = 37,

}

Wait reason for the previous thread. The ordering is important as based on the OldThreadWaitReason definition from the link above. The following are the possible values:

Variants

Executive = 0

FreePage = 1

PageIn = 2

PoolAllocation = 3

DelayExecution = 4

Suspend = 5

UserRequest = 6

WrExecutive = 7

WrFreePage = 8

WrPageIn = 9

WrPoolAllocation = 10

WrDelayExecution = 11

WrSuspended = 12

WrUserRequest = 13

WrEventPair = 14

WrQueue = 15

WrLpcReceiver = 16

WrLpcReply = 17

WrVirtualMemory = 18

WrPageOut = 19

WrRendezVous = 20

WrKeyedEvent = 21

WrTerminated = 22

WrProcessInSwap = 23

WrCpuRateControl = 24

WrCalloutStack = 25

WrKernel = 26

WrResource = 27

WrPushLock = 28

WrMutex = 29

WrQuantumEnd = 30

WrDispatchInt = 31

WrPreempted = 32

WrYieldExecution = 33

WrFastMutex = 34

WrGuardMutex = 35

WrRundown = 36

MaximumWaitReason = 37

Implementations

impl OldThreadWaitReason

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of OldThreadWaitReason.

pub fn from_i32(value: i32) -> Option<OldThreadWaitReason>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a OldThreadWaitReason, or None if value is not a valid variant.

impl OldThreadWaitReason

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.