micromegas_perfetto::protos::gpu_counter_descriptor

Enum MeasureUnit

#[repr(i32)]
pub enum MeasureUnit {

    None = 0,
    Bit = 1,
    Kilobit = 2,
    Megabit = 3,
    Gigabit = 4,
    Terabit = 5,
    Petabit = 6,
    Byte = 7,
    Kilobyte = 8,
    Megabyte = 9,
    Gigabyte = 10,
    Terabyte = 11,
    Petabyte = 12,
    Hertz = 13,
    Kilohertz = 14,
    Megahertz = 15,
    Gigahertz = 16,
    Terahertz = 17,
    Petahertz = 18,
    Nanosecond = 19,
    Microsecond = 20,
    Millisecond = 21,
    Second = 22,
    Minute = 23,
    Hour = 24,
    Vertex = 25,
    Pixel = 26,
    Triangle = 27,
    Primitive = 38,
    Fragment = 39,
    Milliwatt = 28,
    Watt = 29,
    Kilowatt = 30,
    Joule = 31,
    Volt = 32,
    Ampere = 33,
    Celsius = 34,
    Fahrenheit = 35,
    Kelvin = 36,
    Percent = 37,
    Instruction = 40,

}

next id: 41

Variants

None = 0

Bit = 1

Kilobit = 2

Megabit = 3

Gigabit = 4

Terabit = 5

Petabit = 6

Byte = 7

Kilobyte = 8

Megabyte = 9

Gigabyte = 10

Terabyte = 11

Petabyte = 12

Hertz = 13

Kilohertz = 14

Megahertz = 15

Gigahertz = 16

Terahertz = 17

Petahertz = 18

Nanosecond = 19

Microsecond = 20

Millisecond = 21

Second = 22

Minute = 23

Hour = 24

Vertex = 25

Pixel = 26

Triangle = 27

Primitive = 38

Fragment = 39

Milliwatt = 28

Watt = 29

Kilowatt = 30

Joule = 31

Volt = 32

Ampere = 33

Celsius = 34

Fahrenheit = 35

Kelvin = 36

Percent = 37

Values should be out of 100.

Instruction = 40

Implementations

impl MeasureUnit

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of MeasureUnit.

pub fn from_i32(value: i32) -> Option<MeasureUnit>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a MeasureUnit, or None if value is not a valid variant.

impl MeasureUnit

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.