micromegas_perfetto::protos

Struct InputWindowInfoProto

pub struct InputWindowInfoProto {

    pub layout_params_flags: Option<u32>,
    pub layout_params_type: Option<i32>,
    pub frame: Option<RectProto>,
    pub touchable_region: Option<RegionProto>,
    pub surface_inset: Option<i32>,
    pub visible: Option<bool>,
    pub can_receive_keys: Option<bool>,
    pub focusable: Option<bool>,
    pub has_wallpaper: Option<bool>,
    pub global_scale_factor: Option<f32>,
    pub window_x_scale: Option<f32>,
    pub window_y_scale: Option<f32>,
    pub crop_layer_id: Option<i32>,
    pub replace_touchable_region_with_crop: Option<bool>,
    pub touchable_region_crop: Option<RectProto>,
    pub transform: Option<TransformProto>,
    pub input_config: Option<u32>,

}

Fields
layout_params_flags: Option<u32>layout_params_type: Option<i32>frame: Option<RectProto>touchable_region: Option<RegionProto>surface_inset: Option<i32>visible: Option<bool>can_receive_keys: Option<bool>
👎Deprecated
focusable: Option<bool>has_wallpaper: Option<bool>global_scale_factor: Option<f32>window_x_scale: Option<f32>
👎Deprecated
window_y_scale: Option<f32>
👎Deprecated
crop_layer_id: Option<i32>replace_touchable_region_with_crop: Option<bool>touchable_region_crop: Option<RectProto>transform: Option<TransformProto>input_config: Option<u32>
Implementations

impl InputWindowInfoProto

pub fn layout_params_flags(&self) -> u32

Returns the value of layout_params_flags, or the default value if layout_params_flags is unset.

pub fn layout_params_type(&self) -> i32

Returns the value of layout_params_type, or the default value if layout_params_type is unset.

pub fn surface_inset(&self) -> i32

Returns the value of surface_inset, or the default value if surface_inset is unset.

pub fn visible(&self) -> bool

Returns the value of visible, or the default value if visible is unset.

pub fn can_receive_keys(&self) -> bool

Returns the value of can_receive_keys, or the default value if can_receive_keys is unset.

pub fn focusable(&self) -> bool

Returns the value of focusable, or the default value if focusable is unset.

pub fn has_wallpaper(&self) -> bool

Returns the value of has_wallpaper, or the default value if has_wallpaper is unset.

pub fn global_scale_factor(&self) -> f32

Returns the value of global_scale_factor, or the default value if global_scale_factor is unset.

pub fn window_x_scale(&self) -> f32

Returns the value of window_x_scale, or the default value if window_x_scale is unset.

pub fn window_y_scale(&self) -> f32

Returns the value of window_y_scale, or the default value if window_y_scale is unset.

pub fn crop_layer_id(&self) -> i32

Returns the value of crop_layer_id, or the default value if crop_layer_id is unset.

pub fn replace_touchable_region_with_crop(&self) -> bool

Returns the value of replace_touchable_region_with_crop, or the default value if replace_touchable_region_with_crop is unset.

pub fn input_config(&self) -> u32

Returns the value of input_config, or the default value if input_config is unset.