micromegas_perfetto::protos

Struct LayerState

pub struct LayerState {

    pub layer_id: Option<u32>,
    pub what: Option<u64>,
    pub x: Option<f32>,
    pub y: Option<f32>,
    pub z: Option<i32>,
    pub w: Option<u32>,
    pub h: Option<u32>,
    pub layer_stack: Option<u32>,
    pub flags: Option<u32>,
    pub mask: Option<u32>,
    pub matrix: Option<Matrix22>,
    pub corner_radius: Option<f32>,
    pub background_blur_radius: Option<u32>,
    pub parent_id: Option<u32>,
    pub relative_parent_id: Option<u32>,
    pub alpha: Option<f32>,
    pub color: Option<Color3>,
    pub transparent_region: Option<RegionProto>,
    pub transform: Option<u32>,
    pub transform_to_display_inverse: Option<bool>,
    pub crop: Option<RectProto>,
    pub buffer_data: Option<BufferData>,
    pub api: Option<i32>,
    pub has_sideband_stream: Option<bool>,
    pub color_transform: Option<ColorTransformProto>,
    pub blur_regions: Vec<BlurRegion>,
    pub window_info_handle: Option<WindowInfo>,
    pub bg_color_alpha: Option<f32>,
    pub bg_color_dataspace: Option<i32>,
    pub color_space_agnostic: Option<bool>,
    pub shadow_radius: Option<f32>,
    pub frame_rate_selection_priority: Option<i32>,
    pub frame_rate: Option<f32>,
    pub frame_rate_compatibility: Option<i32>,
    pub change_frame_rate_strategy: Option<i32>,
    pub fixed_transform_hint: Option<u32>,
    pub frame_number: Option<u64>,
    pub auto_refresh: Option<bool>,
    pub is_trusted_overlay: Option<bool>,
    pub buffer_crop: Option<RectProto>,
    pub destination_frame: Option<RectProto>,
    pub drop_input_mode: Option<i32>,
    pub trusted_overlay: Option<i32>,

}

Keep insync with layer_state_t

Fields
layer_id: Option<u32>what: Option<u64>x: Option<f32>y: Option<f32>z: Option<i32>w: Option<u32>h: Option<u32>layer_stack: Option<u32>flags: Option<u32>mask: Option<u32>matrix: Option<Matrix22>corner_radius: Option<f32>background_blur_radius: Option<u32>parent_id: Option<u32>relative_parent_id: Option<u32>alpha: Option<f32>color: Option<Color3>transparent_region: Option<RegionProto>transform: Option<u32>transform_to_display_inverse: Option<bool>crop: Option<RectProto>buffer_data: Option<BufferData>api: Option<i32>has_sideband_stream: Option<bool>color_transform: Option<ColorTransformProto>blur_regions: Vec<BlurRegion>window_info_handle: Option<WindowInfo>bg_color_alpha: Option<f32>bg_color_dataspace: Option<i32>color_space_agnostic: Option<bool>shadow_radius: Option<f32>frame_rate_selection_priority: Option<i32>frame_rate: Option<f32>frame_rate_compatibility: Option<i32>change_frame_rate_strategy: Option<i32>fixed_transform_hint: Option<u32>frame_number: Option<u64>auto_refresh: Option<bool>is_trusted_overlay: Option<bool>

unused

buffer_crop: Option<RectProto>destination_frame: Option<RectProto>drop_input_mode: Option<i32>trusted_overlay: Option<i32>
Implementations

impl LayerState

pub fn layer_id(&self) -> u32

Returns the value of layer_id, or the default value if layer_id is unset.

pub fn what(&self) -> u64

Returns the value of what, or the default value if what is unset.

pub fn x(&self) -> f32

Returns the value of x, or the default value if x is unset.

pub fn y(&self) -> f32

Returns the value of y, or the default value if y is unset.

pub fn z(&self) -> i32

Returns the value of z, or the default value if z is unset.

pub fn w(&self) -> u32

Returns the value of w, or the default value if w is unset.

pub fn h(&self) -> u32

Returns the value of h, or the default value if h is unset.

pub fn layer_stack(&self) -> u32

Returns the value of layer_stack, or the default value if layer_stack is unset.

pub fn flags(&self) -> u32

Returns the value of flags, or the default value if flags is unset.

pub fn mask(&self) -> u32

Returns the value of mask, or the default value if mask is unset.

pub fn corner_radius(&self) -> f32

Returns the value of corner_radius, or the default value if corner_radius is unset.

pub fn background_blur_radius(&self) -> u32

Returns the value of background_blur_radius, or the default value if background_blur_radius is unset.

pub fn parent_id(&self) -> u32

Returns the value of parent_id, or the default value if parent_id is unset.

pub fn relative_parent_id(&self) -> u32

Returns the value of relative_parent_id, or the default value if relative_parent_id is unset.

pub fn alpha(&self) -> f32

Returns the value of alpha, or the default value if alpha is unset.

pub fn transform(&self) -> u32

Returns the value of transform, or the default value if transform is unset.

pub fn transform_to_display_inverse(&self) -> bool

Returns the value of transform_to_display_inverse, or the default value if transform_to_display_inverse is unset.

pub fn api(&self) -> i32

Returns the value of api, or the default value if api is unset.

pub fn has_sideband_stream(&self) -> bool

Returns the value of has_sideband_stream, or the default value if has_sideband_stream is unset.

pub fn bg_color_alpha(&self) -> f32

Returns the value of bg_color_alpha, or the default value if bg_color_alpha is unset.

pub fn bg_color_dataspace(&self) -> i32

Returns the value of bg_color_dataspace, or the default value if bg_color_dataspace is unset.

pub fn color_space_agnostic(&self) -> bool

Returns the value of color_space_agnostic, or the default value if color_space_agnostic is unset.

pub fn shadow_radius(&self) -> f32

Returns the value of shadow_radius, or the default value if shadow_radius is unset.

pub fn frame_rate_selection_priority(&self) -> i32

Returns the value of frame_rate_selection_priority, or the default value if frame_rate_selection_priority is unset.

pub fn frame_rate(&self) -> f32

Returns the value of frame_rate, or the default value if frame_rate is unset.

pub fn frame_rate_compatibility(&self) -> i32

Returns the value of frame_rate_compatibility, or the default value if frame_rate_compatibility is unset.

pub fn change_frame_rate_strategy(&self) -> i32

Returns the value of change_frame_rate_strategy, or the default value if change_frame_rate_strategy is unset.

pub fn fixed_transform_hint(&self) -> u32

Returns the value of fixed_transform_hint, or the default value if fixed_transform_hint is unset.

pub fn frame_number(&self) -> u64

Returns the value of frame_number, or the default value if frame_number is unset.

pub fn auto_refresh(&self) -> bool

Returns the value of auto_refresh, or the default value if auto_refresh is unset.

pub fn is_trusted_overlay(&self) -> bool

Returns the value of is_trusted_overlay, or the default value if is_trusted_overlay is unset.

pub fn drop_input_mode(&self) -> DropInputMode

Returns the enum value of drop_input_mode, or the default if the field is unset or set to an invalid enum value.

pub fn set_drop_input_mode(&mut self, value: DropInputMode)

Sets drop_input_mode to the provided enum value.

pub fn trusted_overlay(&self) -> TrustedOverlay

Returns the enum value of trusted_overlay, or the default if the field is unset or set to an invalid enum value.

pub fn set_trusted_overlay(&mut self, value: TrustedOverlay)

Sets trusted_overlay to the provided enum value.