micromegas_perfetto::protos

Module trace_config

Nested message and enum types in TraceConfig.

Modules

buffer_config

Nested message and enum types in BufferConfig.

trace_filter

Nested message and enum types in TraceFilter.

trigger_config

Nested message and enum types in TriggerConfig.

Structs

AndroidReportConfig

Android-only. Not for general use. If set, reports the trace to the Android framework. This field is read by perfetto_cmd, rather than the tracing service. This field must be set when passing the –upload flag to perfetto_cmd.

BufferConfig

BuiltinDataSource

Config for disabling builtin data sources in the tracing service.

CmdTraceStartDelay

If set, delays the start of tracing by a random duration. The duration is chosen from a uniform distribution between the specified minimum and maximum. Note: this delay is implemented by perfetto_cmd not by traced so will not work if you communicate with traced directly over the consumer API. Introduced in Android T.

DataSource

GuardrailOverrides

Contains flags which override the default values of the guardrails inside Perfetto.

IncidentReportConfig

Android-only. Not for general use. If set, saves the trace into an incident. This field is read by perfetto_cmd, rather than the tracing service. This field must be set when passing the –upload flag to perfetto_cmd.

IncrementalStateConfig

Configuration for trace contents that reference earlier trace data. For example, a data source might intern strings, and emit packets containing {interned id : string} pairs. Future packets from that data source can then use the interned ids instead of duplicating the raw string contents. The trace parser will then need to use that interning table to fully interpret the rest of the trace.

ProducerConfig

SessionSemaphore

When non-empty, ensures that for a each semaphore named name at most max_other_session_count`` other sessions (whose value is taken of the minimum of all values specified by this config or any already-running session) can be be running.

StatsdMetadata

Contains statsd-specific metadata about an alert associated with the trace.

TraceFilter

When set applies a post-filter to the trace contents using the filter provided. The filter is applied at ReadBuffers() time and works both in the case of IPC readback and write_into_file. This filter can be generated using tools/proto_filter -s schema.proto -F filter_out.bytes or -T filter_out.escaped_string (for .pbtx). See go/trace-filtering for design.

TriggerConfig

Triggers allow producers to start or stop the tracing session when an event occurs.

Enums

CompressionType

Compress trace with the given method. Best effort.

LockdownModeOperation

StatsdLogging