micromegas_perfetto::protos::trace_packet

Enum Data

pub enum Data {

    ProcessTree(ProcessTree),
    ProcessStats(ProcessStats),
    InodeFileMap(InodeFileMap),
    ChromeEvents(ChromeEventBundle),
    ClockSnapshot(ClockSnapshot),
    SysStats(SysStats),
    TrackEvent(TrackEvent),
    TraceUuid(TraceUuid),
    TraceConfig(TraceConfig),
    FtraceStats(FtraceStats),
    TraceStats(TraceStats),
    ProfilePacket(ProfilePacket),
    StreamingAllocation(StreamingAllocation),
    StreamingFree(StreamingFree),
    Battery(BatteryCounters),
    PowerRails(PowerRails),
    AndroidLog(AndroidLogPacket),
    SystemInfo(SystemInfo),
    Trigger(Trigger),
    ChromeTrigger(ChromeTrigger),
    PackagesList(PackagesList),
    ChromeBenchmarkMetadata(ChromeBenchmarkMetadata),
    PerfettoMetatrace(PerfettoMetatrace),
    ChromeMetadata(ChromeMetadataPacket),
    GpuCounterEvent(GpuCounterEvent),
    GpuRenderStageEvent(GpuRenderStageEvent),
    StreamingProfilePacket(StreamingProfilePacket),
    HeapGraph(HeapGraph),
    GraphicsFrameEvent(GraphicsFrameEvent),
    VulkanMemoryEvent(VulkanMemoryEvent),
    GpuLog(GpuLog),
    VulkanApiEvent(VulkanApiEvent),
    PerfSample(PerfSample),
    CpuInfo(CpuInfo),
    SmapsPacket(SmapsPacket),
    ServiceEvent(TracingServiceEvent),
    InitialDisplayState(InitialDisplayState),
    GpuMemTotalEvent(GpuMemTotalEvent),
    MemoryTrackerSnapshot(MemoryTrackerSnapshot),
    FrameTimelineEvent(FrameTimelineEvent),
    AndroidEnergyEstimationBreakdown(AndroidEnergyEstimationBreakdown),
    UiState(UiState),
    AndroidCameraFrameEvent(AndroidCameraFrameEvent),
    AndroidCameraSessionStats(AndroidCameraSessionStats),
    TranslationTable(TranslationTable),
    AndroidGameInterventionList(AndroidGameInterventionList),
    StatsdAtom(StatsdAtom),
    AndroidSystemProperty(AndroidSystemProperty),
    EntityStateResidency(EntityStateResidency),
    ProfiledFrameSymbols(ProfiledFrameSymbols),
    ModuleSymbols(ModuleSymbols),
    DeobfuscationMapping(DeobfuscationMapping),
    TrackDescriptor(TrackDescriptor),
    ProcessDescriptor(ProcessDescriptor),
    ThreadDescriptor(ThreadDescriptor),
    FtraceEvents(FtraceEventBundle),
    SynchronizationMarker(Vec<u8>),
    CompressedPackets(Vec<u8>),
    ExtensionDescriptor(ExtensionDescriptor),
    NetworkPacket(NetworkPacketEvent),
    NetworkPacketBundle(NetworkPacketBundle),
    TrackEventRangeOfInterest(TrackEventRangeOfInterest),
    SurfaceflingerLayersSnapshot(LayersSnapshotProto),
    SurfaceflingerTransactions(TransactionTraceEntry),
    ShellTransition(ShellTransition),
    ShellHandlerMappings(ShellHandlerMappings),
    ProtologMessage(ProtoLogMessage),
    ProtologViewerConfig(ProtoLogViewerConfig),
    WinscopeExtensions(WinscopeExtensions),
    EtwEvents(EtwTraceEventBundle),
    V8JsCode(V8JsCode),
    V8InternalCode(V8InternalCode),
    V8WasmCode(V8WasmCode),
    V8RegExpCode(V8RegExpCode),
    V8CodeMove(V8CodeMove),
    RemoteClockSync(RemoteClockSync),
    PixelModemEvents(PixelModemEvents),
    PixelModemTokenDatabase(PixelModemTokenDatabase),
    ForTesting(TestEvent),

}

Variants

ProcessTree(ProcessTree)

ProcessStats(ProcessStats)

InodeFileMap(InodeFileMap)

ChromeEvents(ChromeEventBundle)

ClockSnapshot(ClockSnapshot)

SysStats(SysStats)

TrackEvent(TrackEvent)

TraceUuid(TraceUuid)

TraceConfig(TraceConfig)

FtraceStats(FtraceStats)

TraceStats(TraceStats)

ProfilePacket(ProfilePacket)

StreamingAllocation(StreamingAllocation)

StreamingFree(StreamingFree)

Battery(BatteryCounters)

PowerRails(PowerRails)

AndroidLog(AndroidLogPacket)

SystemInfo(SystemInfo)

Trigger(Trigger)

ChromeTrigger(ChromeTrigger)

PackagesList(PackagesList)

ChromeBenchmarkMetadata(ChromeBenchmarkMetadata)

PerfettoMetatrace(PerfettoMetatrace)

ChromeMetadata(ChromeMetadataPacket)

GpuCounterEvent(GpuCounterEvent)

GpuRenderStageEvent(GpuRenderStageEvent)

StreamingProfilePacket(StreamingProfilePacket)

HeapGraph(HeapGraph)

GraphicsFrameEvent(GraphicsFrameEvent)

VulkanMemoryEvent(VulkanMemoryEvent)

GpuLog(GpuLog)

VulkanApiEvent(VulkanApiEvent)

PerfSample(PerfSample)

CpuInfo(CpuInfo)

SmapsPacket(SmapsPacket)

ServiceEvent(TracingServiceEvent)

InitialDisplayState(InitialDisplayState)

GpuMemTotalEvent(GpuMemTotalEvent)

MemoryTrackerSnapshot(MemoryTrackerSnapshot)

FrameTimelineEvent(FrameTimelineEvent)

AndroidEnergyEstimationBreakdown(AndroidEnergyEstimationBreakdown)

UiState(UiState)

AndroidCameraFrameEvent(AndroidCameraFrameEvent)

AndroidCameraSessionStats(AndroidCameraSessionStats)

TranslationTable(TranslationTable)

AndroidGameInterventionList(AndroidGameInterventionList)

StatsdAtom(StatsdAtom)

AndroidSystemProperty(AndroidSystemProperty)

EntityStateResidency(EntityStateResidency)

ProfiledFrameSymbols(ProfiledFrameSymbols)

Only used in profile packets.

ModuleSymbols(ModuleSymbols)

DeobfuscationMapping(DeobfuscationMapping)

TrackDescriptor(TrackDescriptor)

Only used by TrackEvent.

ProcessDescriptor(ProcessDescriptor)

Deprecated, use TrackDescriptor instead.

ThreadDescriptor(ThreadDescriptor)

Deprecated, use TrackDescriptor instead.

FtraceEvents(FtraceEventBundle)

Events from the Linux kernel ftrace infrastructure.

SynchronizationMarker(Vec<u8>)

This field is emitted at periodic intervals (~10s) and contains always the binary representation of the UUID {82477a76-b28d-42ba-81dc-33326d57a079}. This is used to be able to efficiently partition long traces without having to fully parse them.

CompressedPackets(Vec<u8>)

Zero or more proto encoded trace packets compressed using deflate. Each compressed_packets TracePacket (including the two field ids and sizes) should be less than 512KB.

ExtensionDescriptor(ExtensionDescriptor)

Data sources can extend the trace proto with custom extension protos (see docs/design-docs/extensions.md). When they do that, the descriptor of their extension proto descriptor is serialized in this packet. This allows trace_processor to deserialize extended messages using reflection even if the extension proto is not checked in the Perfetto repo.

NetworkPacket(NetworkPacketEvent)

Represents a single packet sent or received by the network.

NetworkPacketBundle(NetworkPacketBundle)

Represents one or more packets sent or received by the network.

TrackEventRangeOfInterest(TrackEventRangeOfInterest)

The “range of interest” for track events. See the message definition comments for more details.

SurfaceflingerLayersSnapshot(LayersSnapshotProto)

Winscope traces

SurfaceflingerTransactions(TransactionTraceEntry)

ShellTransition(ShellTransition)

ShellHandlerMappings(ShellHandlerMappings)

ProtologMessage(ProtoLogMessage)

ProtologViewerConfig(ProtoLogViewerConfig)

WinscopeExtensions(WinscopeExtensions)

EtwEvents(EtwTraceEventBundle)

Events from the Windows etw infrastructure.

V8JsCode(V8JsCode)

V8InternalCode(V8InternalCode)

V8WasmCode(V8WasmCode)

V8RegExpCode(V8RegExpCode)

V8CodeMove(V8CodeMove)

RemoteClockSync(RemoteClockSync)

Clock synchronization with remote machines.

PixelModemEvents(PixelModemEvents)

PixelModemTokenDatabase(PixelModemTokenDatabase)

ForTesting(TestEvent)

This field is only used for testing. In previous versions of this proto this field had the id 268435455 This caused many problems:

protozero decoder does not handle field ids larger than 999.

old versions of protoc produce Java bindings with syntax errors when the field id is large enough.

Implementations

impl Data

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<Data>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.